updating with train_model function and moving around files
