+++ beyondml/tflow/utils/__init__.py
@@ -1,5 +1,4 @@
 """Some utilities to use when building, loading, and training MANN models"""
 
-from .utils import get_custom_objects, mask_model, remove_layer_masks, add_layer_masks, quantize_model, get_task_masking_gradients, mask_task_weights, train_model_iteratively, train_model
-from .ActiveSparsification import ActiveSparsification
+from .utils import get_custom_objects, mask_model, remove_layer_masks, add_layer_masks, quantize_model, get_task_masking_gradients, mask_task_weights, train_model_iteratively, train_model, ActiveSparsification
 from .transformer import build_token_position_embedding_block, build_transformer_block
